requirements.txt
Markdown
python-dotenv
Requests
mdx_truly_sane_lists
markdown_link_attr_modifier
click

